[Silver III] Title: 주유소, Time: 424 ms, Memory: 36204 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -7,20 +7,20 @@ public class Main {
         StringTokenizer st;
         
         int n = Integer.parseInt(br.readLine());
-        int[] dist = new int[n - 1];
-        int[] gas = new int[n];
+        long[] dist = new long[n - 1];
+        long[] gas = new long[n];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n - 1; i++) {
-            dist[i] = Integer.parseInt(st.nextToken());
+            dist[i] = Long.parseLong(st.nextToken());
         }
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n; i++) {
-            gas[i] = Integer.parseInt(st.nextToken());
+            gas[i] = Long.parseLong(st.nextToken());
         }
         
-        int cost = gas[0] * dist[0];
-        int curr_dist = 0;
-        int target = 0;
+        long cost = gas[0] * dist[0];
+        long curr_dist = 0;
+        long target = 0;
         if (gas[0] < gas[1]) target = gas[0];
         else target = gas[1];
         for (int i = 2; i < n; i++) {
